UI/tray/misc: a11y names, Delete scope, tray restore, themed zebra, small guards.
- Localized AutomationProperties.Name on all six speed-limit flyout controls (N14).
- Scope the Delete accelerator to the downloads table (N15): window-global, it stole
  the key from every TextBox/NumberBox while a row was selected.
- Second-instance activation routes through the tray restore (N16): it runs under the
  _restoring guard so the minimize-watcher can''t re-hide the window mid-restore; the
  restore itself now un-minimizes only when actually minimized, so it no longer
  un-maximizes a maximized window.
- Resolve the zebra-row brush from page theme dictionaries by ActualTheme (N27):
  the app-level lookup ignored the in-app forced theme; no stripe in High Contrast.
- Compare the bencode string length in long arithmetic (N19): an int overflow slipped
  past the bounds check into a ~2 GB allocation.
- Generation token on speed-limit edits (N20): an edit made while the flyout''s
  current-limit fetch is in flight is no longer visually reverted by the stale values.
- Set the window icon by absolute path (U2): a relative path resolves against the
  process current directory in the portable build.

+++ Aria2Gui/Services/TorrentParser.cs
@@ -117,7 +117,9 @@ public static class TorrentParser
                 throw new InvalidDataException(Aria2Gui.Helpers.L.Get("TorrentErrBadString"));
             if (!int.TryParse(Encoding.ASCII.GetString(data, pos, colon - pos), out int length))
                 throw new InvalidDataException(Aria2Gui.Helpers.L.Get("TorrentErrBadString"));
-            if (length < 0 || colon + 1 + length > data.Length)
+            // Compare in long: a declared length near int.MaxValue would wrap the int sum
+            // negative, slip past this check, and hit a ~2 GB allocation below instead.
+            if (length < 0 || (long)colon + 1 + length > data.Length)
                 throw new InvalidDataException(Aria2Gui.Helpers.L.Get("TorrentErrStringOOB"));
             byte[] bytes = new byte[length];
             Array.Copy(data, colon + 1, bytes, 0, length);
